Renaming lp-token->lp-tokens, it handles more than 1 lp token
